:hammer: 층 정보 조회 api의 ResponseDto인터페이스 위치 변경
