+++ src/main/java/com/vincent/domain/building/controller/BuildingController.java
@@ -8,8 +8,8 @@ import com.vincent.domain.building.converter.BuildingConverter;
 import com.vincent.domain.building.entity.Building;
 import com.vincent.domain.building.entity.Floor;
 import com.vincent.domain.building.entity.Space;
-import com.vincent.domain.building.repository.FloorRepository.FloorInfoProjection;
-import com.vincent.domain.building.repository.SpaceRepository.SpaceInfoProjection;
+import com.vincent.domain.building.controller.dto.BuildingResponseDto.FloorInfoProjection;
+import com.vincent.domain.building.controller.dto.BuildingResponseDto.SpaceInfoProjection;
 import com.vincent.domain.building.service.BuildingService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
