+++ src/main/java/com/vincent/domain/building/converter/BuildingConverter.java
@@ -10,8 +10,8 @@ import com.vincent.domain.building.entity.Building;
 import com.vincent.domain.building.entity.Floor;
 import com.vincent.domain.building.entity.Space;
 import com.vincent.domain.building.repository.FloorRepository;
-import com.vincent.domain.building.repository.FloorRepository.FloorInfoProjection;
-import com.vincent.domain.building.repository.SpaceRepository.SpaceInfoProjection;
+import com.vincent.domain.building.controller.dto.BuildingResponseDto.FloorInfoProjection;
+import com.vincent.domain.building.controller.dto.BuildingResponseDto.SpaceInfoProjection;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.data.domain.Page;
