+++ src/main/java/com/vincent/domain/building/service/BuildingService.java
@@ -10,9 +10,9 @@ import com.vincent.domain.building.entity.Floor;
 import com.vincent.domain.building.entity.Space;
 import com.vincent.domain.building.repository.BuildingRepository;
 import com.vincent.domain.building.repository.FloorRepository;
-import com.vincent.domain.building.repository.FloorRepository.FloorInfoProjection;
+import com.vincent.domain.building.controller.dto.BuildingResponseDto.FloorInfoProjection;
 import com.vincent.domain.building.repository.SpaceRepository;
-import com.vincent.domain.building.repository.SpaceRepository.SpaceInfoProjection;
+import com.vincent.domain.building.controller.dto.BuildingResponseDto.SpaceInfoProjection;
 import com.vincent.domain.member.entity.Member;
 import com.vincent.domain.socket.entity.Socket;
 import com.vincent.domain.socket.repository.SocketRepository;
